brownie/requirements.txt
eth-brownie>=1.10.0,<2.0.0
vyper>=0.2.3,<0.3.0
eth-account==0.5.2
python-dotenv==0.13.0